wms-system/backend/requirements.txt
# Django核心
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7

# 工具库
python-dotenv==1.0.0
Pillow==9.5.0
python-jose==3.3.0
pytz==2023.3

# 开发工具
django-debug-toolbar==4.2.0
